circle collision fix

+++ Gizmoball/src/model/CircleBumper.java
@@ -40,7 +40,7 @@ public class CircleBumper extends Gizmo {
     @Override
     public List<Circle> getCorners(){
         List<Circle> cs = new ArrayList<>();
-        cs.add(new Circle((int)(super.getPos().x()+0.5), super.getPos().y()+0.5, 0.5));
+        cs.add(new Circle(super.getPos().x()+0.5, super.getPos().y()+0.5, 0.5));
         return cs;
     }
 }
